requirements.txt
asgiref==3.5.2
astroid==2.12.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==3.2.16
django-allauth==0.51.0
django-storages==1.13.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
oauthlib==3.2.1
openapi-codec==1.3.2
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.5.0
pylint-plugin-utils==0.7
pyrsistent==0.18.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
sqlparse==0.4.3
static3==0.7.0
tomli==2.0.1
tomlkit==0.11.5
types-cryptography==3.3.23
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
